bridge_hub_westend_runtime

Module bridge_to_rococo_config

Bridge definitions used on BridgeHub with the Westend flavor.

Modules

migration

Contains the migration for the AssetHubWestend<>AssetHubRococo bridge.

Structs

BridgeDeposit

BridgeHubRococoLocation

BridgeWestendToRococoMessagesPalletInstance

CongestionManager

Implementation of bp_xcm_bridge_hub::LocalXcmChannelManager for congestion management.

DeliveryRewardInBalance

MaxRococoParaHeadDataSize

ParachainHeadsToKeep

PriorityBoostPerMessage

PriorityBoostPerParachainHeader

PriorityBoostPerRelayHeader

RelayChainHeadersToKeep

RococoBridgeParachainPalletName

RococoGlobalConsensusNetwork

RococoGlobalConsensusNetworkLocation

StrOnBridgeHubWestendRefundBridgeHubRococoMessages

Type Aliases

BridgeGrandpaRococoInstance

Add GRANDPA bridge pallet to track Rococo relay chain.

BridgeParachainRococoInstance

Add parachain bridge pallet to track Rococo BridgeHub parachain

FromRococoBridgeHubMessagesProof

Proof of messages, coming from Rococo.

OnBridgeHubWestendRefundBridgeHubRococoMessages

Transaction extension that refunds relayers that are delivering messages from the Rococo parachain.

ToRococoBridgeHubMessagesDeliveryProof

Messages delivery proof for Rococo Bridge Hub -> Westend Bridge Hub messages.

WithBridgeHubRococoMessagesInstance

Add XCM messages support for BridgeHubWestend to support Westend->Rococo XCM messages

XcmOverBridgeHubRococoInstance

Add support for the export and dispatch of XCM programs.