Module bridge_hub_westend_runtime::xcm_config

Structs

MaxAssetsIntoHolding

MaxInstructions

ParentOrParentsPlurality

ParentOrSiblings

RelayChainOrigin

RelayNetwork

RelayTreasuryLocation

SafeCallFilter

A call filter for the XCM Transact instruction. This is a temporary measure until we properly account for proof size weights.

SystemParachains

TreasuryAccount

UniversalLocation

WestendLocation

XcmConfig

Type Definitions

Barrier

CurrencyTransactor

Means for transacting the native currency on this chain.

LocalOriginToLocation

Converts a local signed origin into an XCM multilocation. Forms the basis for local origins sending/executing XCMs.

LocationToAccountId

Type for specifying how a MultiLocation can be converted into an AccountId. This is used when determining ownership of accounts for asset transacting and when attempting to use XCM Transact in order to determine the dispatch Origin.

PriceForParentDelivery

TrustedTeleporters

Cases where a remote origin is accepted as trusted Teleporter for a given asset:

WaivedLocations

Locations that will not be charged fees in the executor, either execution or delivery. We only waive fees for system functions, which these locations represent.

XcmOriginToTransactDispatchOrigin

This is the type we use to convert an (incoming) XCM origin into a local Origin instance, ready for dispatching a transaction with Xcm’s Transact. There is an OriginKind which can biases the kind of local Origin it will become.

XcmRouter

The means for routing XCM messages which are not for local execution into the right message queues.